+++ modules/uhsg_office_hours/src/OfficeHoursService.php
@@ -7,6 +7,8 @@ use Drupal\Core\Cache\CacheBackendInterface;
 use Drupal\Core\Config\ConfigFactory;
 use Drupal\Core\Language\LanguageManagerInterface;
 use Drupal\Core\Logger\LoggerChannel;
+use Drupal\Core\Messenger\MessengerInterface;
+use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\uhsg_active_degree_programme\ActiveDegreeProgrammeService;
 use GuzzleHttp\Client;
 use GuzzleHttp\RequestOptions;
@@ -14,15 +16,13 @@ use Psr\Http\Message\ResponseInterface;
 
 class OfficeHoursService {
 
-  // 1 minute.
-  const CACHE_EXPIRE_SECONDS = 60;
+  use StringTranslationTrait;
+
   const CACHE_KEY_PREFIX = 'uhsg-office-hours-';
   const CONFIG_NAME = 'uhsg_office_hours.config';
   const CONFIG_API_BASE_URL = 'api_base_url';
   const CONFIG_API_PATH = 'api_path';
-  const CONNECT_TIMEOUT_SECONDS = 2;
   const LANGUAGE_UNDEFINED = 'undefined';
-  const REQUEST_TIMEOUT_SECONDS = 2;
 
   /** @var \Drupal\Core\Cache\CacheBackendInterface*/
   protected $cache;
@@ -36,11 +36,14 @@ class OfficeHoursService {
   /** @var \Drupal\Core\Config\ConfigFactory*/
   protected $configFactory;
 
-  /** @var \Drupal\Core\Language\LanguageManagerInterface */
+  /** @var \Drupal\Core\Language\LanguageManagerInterface*/
   protected $languageManager;
 
   /** @var \Drupal\Core\Logger\LoggerChannel*/
   protected $logger;
+
+  /** @var \Drupal\Core\Messenger\MessengerInterface*/
+  protected $messenger;
 
   /** @var \Drupal\Component\Datetime\TimeInterface*/
   protected $time;
@@ -58,7 +61,8 @@ class OfficeHoursService {
     LoggerChannel $logger,
     TimeInterface $time,
     ActiveDegreeProgrammeService $activeDegreeProgrammeService,
-    LanguageManagerInterface $languageManager) {
+    LanguageManagerInterface $languageManager,
+    MessengerInterface $messenger) {
 
     $this->cache = $cache;
     $this->client = $client;
@@ -67,6 +71,7 @@ class OfficeHoursService {
     $this->time = $time;
     $this->activeDegreeProgrammeService = $activeDegreeProgrammeService;
     $this->languageManager = $languageManager;
+    $this->messenger = $messenger;
   }
 
   /**
@@ -90,6 +95,7 @@ class OfficeHoursService {
         }
         catch (\Exception $e) {
           $this->logger->error($e->getMessage());
+          $this->messenger->addError($this->t('The office hours cannot be displayed. Please try again later.'));
         }
       }
     }
@@ -138,8 +144,8 @@ class OfficeHoursService {
    */
   private function getRequestOptions() {
     return [
-      RequestOptions::CONNECT_TIMEOUT => self::CONNECT_TIMEOUT_SECONDS,
-      RequestOptions::TIMEOUT => self::REQUEST_TIMEOUT_SECONDS
+      RequestOptions::CONNECT_TIMEOUT => $this->config->get('connect_timeout'),
+      RequestOptions::TIMEOUT => $this->config->get('request_timeout')
     ];
   }
 
@@ -287,7 +293,7 @@ class OfficeHoursService {
    * @return int
    */
   private function getCacheExpireTimestamp() {
-    return $this->time->getRequestTime() + self::CACHE_EXPIRE_SECONDS;
+    return $this->time->getRequestTime() + $this->config->get('cache_expire');
   }
 
 }